+++ testing/test_cases/argument_parsing_test_cases.py
@@ -6,25 +6,15 @@ from flake8_annotations.ast_walker import Argument
 from flake8_annotations.enums import AnnotationType
 
 
-class ArgumentTestCase(NamedTuple):
-    """
-    Helper container for Argument parsing test cases.
-
-    The `py38_only` flag may be optionally specified for use skipping test cases that will fail for
-    Python versions less than 3.8
-    """
-
+class ArgumentTestCase(NamedTuple):  # noqa: D101
     src: str
     args: Tuple[Argument, ...]
-    py38_only: bool = False
 
 
 # Note: For testing purposes, lineno and col_offset are ignored so these are set to dummy values
 # using partial objects
 untyped_arg = partial(Argument, lineno=0, col_offset=0, has_type_annotation=False)
-typed_arg = partial(
-    Argument, lineno=0, col_offset=0, has_type_annotation=True, has_3107_annotation=True
-)
+typed_arg = partial(Argument, lineno=0, col_offset=0, has_type_annotation=True)
 
 argument_test_cases = {
     "all_args_untyped": ArgumentTestCase(
@@ -68,7 +58,6 @@ argument_test_cases = {
             untyped_arg(argname="posonlyarg", annotation_type=AnnotationType.POSONLYARGS),
             typed_arg(argname="return", annotation_type=AnnotationType.RETURN),
         ),
-        py38_only=True,
     ),
     "posonly_arg_typed": ArgumentTestCase(
         src=dedent(
@@ -81,7 +70,6 @@ argument_test_cases = {
             typed_arg(argname="posonlyarg", annotation_type=AnnotationType.POSONLYARGS),
             typed_arg(argname="return", annotation_type=AnnotationType.RETURN),
         ),
-        py38_only=True,
     ),
     "posonly_and_arg_args": ArgumentTestCase(
         src=dedent(
@@ -95,6 +83,5 @@ argument_test_cases = {
             typed_arg(argname="bar", annotation_type=AnnotationType.ARGS),
             typed_arg(argname="return", annotation_type=AnnotationType.RETURN),
         ),
-        py38_only=True,
     ),
 }